integrated elastic, kibana and docker compose

+++ NotificationService/src/main/java/com/notificationservice/service/NotificationService.java
@@ -11,8 +11,8 @@ import org.springframework.mail.MailSender;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class NotificationService {
@@ -37,31 +37,35 @@ public class NotificationService {
         notificationsRepository.save(notification);
     }
     public boolean hasBeenSent(NotificationDTO notification){
-        return notificationsRepository.findByRecipientIdAndAndBookingStatusAndStatus(notification.getRecipientId(),
+        return notificationsRepository.findByRecipientIdAndBookingStatusAndStatus(notification.getRecipientId(),
                         notification.getBookingStatus(),
                         NotificationStatus.DELIVERED)
                 .isPresent();
     }
 
-    @Scheduled(cron = "0 * * * * ?")
+    @Scheduled(cron = "0 0 0 * * ?")
     public void safeDeleteDeliveredNotifications(){
         notificationsRepository.deleteByStatus(NotificationStatus.DELIVERED);
     }
 
     @Scheduled(cron="0 * * * * ?")
-    public void resendFailedNotifications(){
+    public void resendFailedNotifications() {
         List<Notification> failedNotifications = notificationsRepository.findByStatus(NotificationStatus.FAILED);
-        List<NotificationDTO> failedNotificationsDTO = notificationMapper.toDTOList(failedNotifications);
-        for(NotificationDTO notification: failedNotificationsDTO){
-            notification.setRetryCount(notification.getRetryCount() + 1);
-                if(notification.getRetryCount() <= 3)
+        List<NotificationDTO> failedNotificationsDTO = failedNotifications.stream()
+                        .map(notificationMapper::toDTO)
+                        .toList();
+        if (!failedNotificationsDTO.isEmpty()){
+            for (NotificationDTO notification : failedNotificationsDTO) {
+                notification.setRetryCount(notification.getRetryCount() + 1);
+                if (notification.getRetryCount() <= 3) {
                     kafkaProducerService.sendMessage(notification);
-                else{
+                } else {
                     NotificationDTO failedSendingNotification = bookingHandlingService.sendFailedNotificationReceiving(notification.getEmail(), notification.getRecipientId());
                     SimpleMailMessage simpleMailMessage = simpleMailMessageMapper.toSimpleMailMessage(notificationMapper.toEntity(failedSendingNotification));
                     mailSender.send(simpleMailMessage);
                     notificationsRepository.deleteById(notification.getId());
                 }
             }
+    }
     }
 }
